+++ src/nesterov.h
@@ -27,15 +27,15 @@
 namespace opt {
     class NAGoptimizer {
       private:
-        float lipschitz_;
-        Array<float> t_;
-        Array<float> z_;
+        double lipschitz_;
+        Array<double> t_;
+        Array<double> z_;
 
       public:
-        NAGoptimizer(int num_pixels, int num_projs, const Array<float> & angles, float cen):
+        NAGoptimizer(int num_pixels, int num_projs, const Array<double> & angles, double cen):
             t_(num_pixels, num_pixels), z_(num_pixels, num_pixels) {
 
-            Array<float> x(num_pixels, num_pixels, 1);
+            Array<double> x(num_pixels, num_pixels, 1);
             auto y = forward(x, angles, cen);
             auto g = backward(y, angles, cen);
             lipschitz_ = 1./g.max();
@@ -45,9 +45,9 @@ namespace opt {
         void update(Array<T> &recon, Array<T> &gradient) {
             #pragma omp parallel for
             for (int i = 0; i < recon.size(); i++) {
-                float z_new = recon[i] - lipschitz_ * gradient[i];
-                float t_new = 0.5 * (1 + std::sqrt(1 + 4 * t_[i] * t_[i]));
-                float l = (1 - t_[i]) / t_new;
+                double z_new = recon[i] - lipschitz_ * gradient[i];
+                double t_new = 0.5 * (1 + std::sqrt(1 + 4 * t_[i] * t_[i]));
+                double l = (1 - t_[i]) / t_new;
                 recon[i] = (1 - l) * z_new + l * z_[i];
                 t_[i] = t_new;
                 z_[i] = z_new;
